[database:dump] Keep the overwritten file

+++ src/Command/Database/DumpCommand.php
@@ -137,7 +137,9 @@ class DumpCommand extends Command
 													$file)
 									)
 							);
-							unlink($file);
+							if($resultFile != $file) {
+								unlink($file);
+							}
 						}
 
             $io->success(
